Point the main config file to an icinga-specific file
There are some configurations in that file that are icinga-specific, so
we need to diverge from the nagios configuration. If we use the same
configuration file as the one used by nagios, the icinga service will
just not start.

There actually was already a default icinga configuration file so let's
just point the resource to the right sources so that this
icinga-specific default file will be used.