get auto purging working again
That purging resources work correctly, the config files must be at
the location that the nagios providers have internally, which is
at /etc/nagios/<name>.cfg
To ease the config file we used the cfg_dir option so far, which we
can't anymore and hence this change requires adoption of your site
modules. But auto-purging resources will then finally work.