stop defining an alias by default for hosts
Since the "alias" attribute is special for puppet in that it needs to be
a string that's unique throughout all of the resources on a given host's
catalogue, we shouldn't always define an alias unless users really want
them.

This alias might (and did for some ppl) create a conflict when using a
class that is named the same as $::hostname.