files managed for apache on debian break usage through main class
Currently the nagios::apache class manages some files for debian. This
means that when using the main class for setting up nagios with apache
(which is the default!) lacks those files and the setup doesn't work.

Split out management of those files into another class to avoid making
the code in the main class too encumbered and make the main class
include that (instead of nagios::apache). nagios::apache will include
the main class anyway.

Adjust README file so that all cases of http management are done through
the "nagios" class.